Enabled PTS by default, added the option to set the pts profile from commandline, defaulting to FAST. Set up github actions to use different profiles and modes.
The build verification will use `REMAINING_TESTS` if it's started on the main branch and `RELEVANT_TESTS` otherwise. If it's a PR targeting the main branch, it will use the `CONSERVATIVE` profile, otherwise (PR to another branch) it will use the `STANDARD` profile. `FAST` profile is not used on the CI.