Use provider API so that version.json is configuration cache input
Previously when making changes to the version.json file would not cause the configuration cache to re-execute since the contents of the versions.json file were not inputs to configuration cache. As an example, if you bumped the version from  8.4.0-beta01 to 8.4.0-beta02, the task name would still be `beta01` since the configuration cache would re-run.

This also bumps the supported version to 8.4.0-beta02