Adjust Renovate config (#1750)
- Always allow pre-release in `latestKnownAgpVersion` property
    
Default behavior would be to only allow it while it's already unstable
(currently the case).

- Disable dependency dashboard (#1746)

 - Migrate deprecated setting
    
Assuming this is the main change proposed by Renovate in #1749. It's a
noisy diff there due to syntax changes.

- Group all AGP version changes in single PR

This should prevent the check failure that's currently affecting #1747
and #1748. Making them a single PR should fix that error.

These changes were tested on a fork. An example PR is
gabrielfeo/android-cache-fix-gradle-plugin#12). A [help][1] build on
that branch confirms that the new PR should pass the current error.

[1]: https://ge.solutions-team.gradle.com/s/hvjxrb36dfoja