Update Renovate config to avoid JSON5 (#1753)
In #1742, Renovate was set up with a more usual standard `regex` matcher
configuration, which is to search for versions in lines annotated with a
comment `// renovate: ...`. However, the `versions.json` file is parsed
by several tools which don't support JSON5 or any form of comments,
namely Groovy's `json` package and `jq`. Groovy `json` required using a
lenient mode (#1742), which may suppress useful validations, while `jq`
required stripping the comments beforehand (#1752). Moving away from
JSON5 would be a simpler solution.

Revert the `versions.json` back from JSON5 to plain JSON and change the
Renovate config to match the version properties based on the fact that
their values are lists, e.g. `"8.8.0": [`. Matching the version in
`gradle.properties` is also simplified to no longer require a comment.

Tested with a [dry-run of gradle/renovate-agent][2] against this repo
and the Renovate GitHub app against a fork, which created an [example
PR][1].

[1]:
https://github.com/gabrielfeo/android-cache-fix-gradle-plugin/pull/14#issue-2787066376
[2]:
https://github.com/gradle/renovate-agent/actions/runs/12771500371/job/35598810573