Clarify distinction between "supported" and "tested" AGP versions (#1779)
`versions.json` values are currently referred to as "supported" versions throughout project source code. Even though it _is_ used to determine the minimum supported AGP version, it's not a comprehensive list of versions that the plugin supports, which may lead to confusion. For instance, the plugin supports a variety of patch versions beyond those. Mainly, the `versions.json` file determines which AGP and Gradle versions the plugin will be _**tested**_ against, not the versions it supports. 

- Change references to AGP versions throughout project source from "supported" to "tested" versions, which more accurately describes the usage of that file
- Move `versions.json` to test resources. It was only needed in the main source set because the minimum AGP version was based on its content. As the minimum is changed very infrequently, it can become hardcoded, allowing the JSON and surrounding code to be moved into the test source set.
- Update Renovate PR titles and commit subjects to a clearer "Update _tested_ AGP versions"

One property keeps "supported" in the name: the `earliestSupportedAndroidVersion` is the minimum AGP version the plugin is _tested_ against. While the minimum _supported_ version is now hardcoded, there's a test that asserts it corresponds to the minimum tested version. No changes are made to the project docs.