Fix (self-hosted) Renovate AGP versions bumping (#1793)
Renovate cannot detect the versions.json file since #1779. #1792 was an
attempt to fix it by removing `config:recommended`:

>The cause is the use of the [`:ignoreModulesAndTests` preset][1], which
is implicitly added by our use of the [`config:recommended` preset][2].
Presets are a mechanism for reuse in Renovate config files. This preset
(as the name indicates) causes any files inside `**/test/**` to be
ignored. When `versions.json` was moved to the test source set in #1779,
it became ignored.

However, [our self-hosted Renovate agent adds `config:recommended` on
its own][3]. While the fix was effective when testing on a fork, it's
ineffective with our Renovate agent.

Fix by adding `ignorePresets`, which can ignore the
`:ignoreModulesAndTests` preset that is coming from the agents config
(preferable than changing the agent, which would affect other projects).
The previously removed `config:recommended` is added back.

[1]: https://docs.renovatebot.com/presets-default/#ignoremodulesandtests
[2]: https://docs.renovatebot.com/presets-config/#configrecommended
[3]:
https://github.com/gradle/renovate-agent/blob/main/.github/workflows/renovate_agent_config.js#L5