Using new configureJavaCompileTask in new variant api (#1894)
AGP 9.0.0-alpha06 introduces the new `configureJavaCompileTask` API,
which provides a dedicated configuration hook for the Java compile task.
This allows us to apply our workaround directly during task
configuration.

This only applies to Application and Library variants (`onVariants`),
and not to unit test or Android test variants (`beforeVariants`). We’ll
create a new issue to follow up with the Android team.