Manual update AGP 9.0.0 alpha13 (#1910)
Renovate PR failed with:
```
> android/databinding/parser/XMLParser$ElementContext has been compiled by a more recent version of the Java Runtime (class file version 65.0), this version of the Java Runtime only recognizes class file versions up to 61.0	
```
The issue can't be reproduced for the same generated project locally and
with the same GHA configuration:
https://github.com/cdsap/9.0.0-alpha13Issue 
Build Scan: https://ge.solutions-team.gradle.com/s/zdbg5wlwyo4ia 

This issue is similar to what we have found in the past with some alpha
versions and databinding:
* https://issuetracker.google.com/issues/279710208
* https://issuetracker.google.com/issues/241287607

Google in one of the issues mentioned:
```
Databinding causing this is just a special case of using a library that was compiled with a newer Kotlin compiler than the one used in the build. The same issue happens with javac when trying to consume jvm bytecode newer than the version of javac you're using.
```

This PR introduces new logic to ignore DataBinding for AGP 9 tests,
based on the trend observed in modern Android projects and the
decreasing usage of this feature.

The only reason we are keeping DataBinding in the generated projects is
for legacy workarounds, which will be removed in the next major version
(along with DataBinding itself).

---------

Signed-off-by: Renovate Bot for Gradle <bot-gh-dv-shared+renovate@gradle.com>
Co-authored-by: Renovate Bot for Gradle <bot-gh-dv-shared+renovate@gradle.com>